pyproject.toml
[tool.poetry]
name = "markovapi"
version = "0.1.0"
description = "API for MarkovBot https://github.com/dmondo/markovbot"
authors = ["dmondo <dannysmondo@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/dmondo/MarkovAPI"
repository = "https://github.com/dmondo/MarkovAPI"
keywords = ["markov", "model", "API", "machine learning"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"

[tool.poetry.dev-dependencies]
pytest-mock = "^3.5.1"
coverage = {version = "^5.3.1", extras = ["toml"]}
pytest-cov = "^2.10.1"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^20.11.1"
safety = "^1.10.3"
mypy = "^0.790"
pytype = {version = "^2021.1.14", python = "3.7"}
typeguard = "^2.10.0"
flake8-annotations = "^2.5.0"
flake8-docstrings = "^1.5.0"
darglint = "^1.5.8"
Sphinx = "^3.4.3"
sphinx-autodoc-typehints = "^1.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 100
